pyproject.toml
[project]
authors = [{name = "digital-idiot", email = "abhisek.maiti@systemiq.earth"}]
description = "Multi-agent CV optimization system using Google ADK"
dependencies = ["google-adk>=1.14.1", "google-genai>=1.52.0", "python-dotenv>=1.2.1", "tika>=3.1.0"]
name = "CVFormatter"
requires-python = ">= 3.11"
version = "0.1.0"

[build-system]
build-backend = "hatchling.build"
requires = ["hatchling"]

[tool.pixi.workspace]
channels = ["conda-forge"]
platforms = ["osx-arm64"]

[tool.pixi.pypi-dependencies]
cvformatter = { path = ".", editable = true }

[tool.pixi.tasks]

[tool.pixi.dependencies]
python = ">=3.14.0,<3.15"
google-genai = ">=1.52.0,<2"
python-dotenv = ">=1.2.1,<2"
tika = ">=3.1.0,<4"
